server.socket/src/main/java/com/likhachova/socket/lessons/Server.java
package com.likhachova.socket.lessons;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

import java.io.*;
import java.net.ServerSocket;
import java.net.Socket;

public class Server {

    private static final Logger LOG = LoggerFactory.getLogger(Server.class);

    private ServerSocket serverSocket;
    private Socket clientSocket;
    private PrintWriter out;
    private BufferedReader in;

    public void start(int port) throws IOException {
        serverSocket = new ServerSocket(port);
        clientSocket = serverSocket.accept();
        out = new PrintWriter(clientSocket.getOutputStream(), true);
        in = new BufferedReader(new InputStreamReader(clientSocket.getInputStream()));
        String content = in.readLine();
        if("echo".equals(content)) {
            out.println("echo client");
        }
        else {
            out.println("unrecognised content");
        }
        stop();
    }

    public void stop() throws IOException {
        in.close();
        out.close();
        clientSocket.close();
        serverSocket.close();
    }

    public static void main(String[] args) {
        Server server = new Server();
        try {
            server.start(3000);
        }
        catch(IOException e) {
            LOG.debug("log start server" + e.getMessage());
            throw new RuntimeException(e);
        }
    }
}
